+++ common/src/main/java/com/team/common/feature_api/navigation_constant/NavigationConstant.kt
@@ -21,4 +21,5 @@ object NavigationRouteConstant {
 
     const val mypageNestedRoute = "mypage_nested_route"
     const val mypageSceneRoute = "mypage_scene_route"
+    const val mypageModifyUserSceneRoute = "mypage_modify_user_scene_route"
 }
