+++ presentation/feature/mypage/src/main/java/com/presentation/mypage/scene/modify_user/ModifyUserViewModel.kt
@@ -1,0 +1,14 @@
+package com.presentation.mypage.scene.modify_user
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+
+@HiltViewModel
+class ModifyUserViewModel @Inject constructor(
+) : ViewModel() {
+    enum class Action {
+
+    }
+}
